gui/requirements-gui.txt
# GUI Application Dependencies
# PySide6 for windowed GUI
PySide6>=6.6.0

# YAML configuration parsing
PyYAML>=6.0
